+++ Assets/Prefabs/PlayerMissle.prefab
@@ -179,4 +179,4 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
-  damage: 1
+  damage: 0
