+++ Assets/Scripts/PlayerLevelManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1a3715b2761d5149849a0242698d213
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
