+++ Assets/Scripts/Store.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d2cef6d15505839438e2a3891efce2f5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
