+++ Assets/Scripts/Weapons/Bullet.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -6,7 +7,8 @@ public class Bullet : MonoBehaviour
 {
     private Vector3 _direction;
     [SerializeField] private float _speed = 10;
-    
+    private const int NUMFIND = 3;
+    [SerializeField] private float _bulletRange = 10;
     public float Speed
     {
         set => _speed = value;
@@ -17,8 +19,28 @@ public class Bullet : MonoBehaviour
         set => _direction = value;
     }
 
+    private void Start()
+    {
+        _direction = AimAtEnemy();
+        transform.rotation = Quaternion.LookRotation(Vector3.forward, _direction); 
+    }
+
     void FixedUpdate()
     {
         transform.position += _direction * _speed;
+    }
+    private Vector3 AimAtEnemy()
+    {
+        var results = new Collider2D[NUMFIND];
+        Physics2D.OverlapCircleNonAlloc(transform.position, _bulletRange, results);
+
+        foreach (var result in results)
+        {
+            if (result!= null && result.CompareTag("Enemy"))
+            {
+                return (result.transform.position - transform.position).normalized;
+            }
+        }
+        return new Vector3(1,0,0);
     }
 }
