+++ Assets/Scripts/Weapons/WeaponFire.cs
@@ -10,11 +10,12 @@ public class WeaponFire : MonoBehaviour
 {
     // Start is called before the first frame update
     
-    [SerializeField] private float _coolDown = 0.6f;
-    [SerializeField] private float _weaponRange = 10;
+    [SerializeField] private float _coolDown;
+
     [SerializeField] private bool fireEnabled = true;
     [SerializeField] private GameObject bullet;
-    private const int NUMFIND = 3;
+    [SerializeField] private UnityEvent FireEvent;
+
     private Vector3 _direction;
     
     private void Start()
@@ -24,28 +25,7 @@ public class WeaponFire : MonoBehaviour
 
     public void FireBullet()
     {
-        _direction = AimAtEnemy();
-        if (_direction == Vector3.zero)
-        {
-            _direction = new Vector3(1, 0, 0);
-        }
-
-        Instantiate(bullet, transform.position, Quaternion.LookRotation(Vector3.forward, _direction));
-        bullet
-    }
-    
-    private Vector3 AimAtEnemy()
-    {
-        var results = new Collider2D[NUMFIND];
-        Physics2D.OverlapCircleNonAlloc(transform.position, _weaponRange, results);
-
-        foreach (var result in results)
-        {
-            if (result!= null && result.CompareTag("Enemy"))
-            {
-                return (result.transform.position - transform.position).normalized;
-            }
-        }
-        return new Vector3(1,0,0);
+        Instantiate(bullet, transform.position, Quaternion.identity);
+        FireEvent.Invoke();
     }
 }
